move tests out of AnalyserIntegrationTest
