+++ tests/PHPStan/Analyser/NodeScopeResolverTest.php
@@ -1327,6 +1327,10 @@ class NodeScopeResolverTest extends TypeInferenceTestCase
 
 		yield from $this->gatherAssertTypes(__DIR__ . '/data/bug-7915.php');
 		yield from $this->gatherAssertTypes(__DIR__ . '/data/bug-9714.php');
+		yield from $this->gatherAssertTypes(__DIR__ . '/data/bug-9105.php');
+		yield from $this->gatherAssertTypes(__DIR__ . '/data/bug-5172.php');
+		yield from $this->gatherAssertTypes(__DIR__ . '/data/bug-9293.php');
+		yield from $this->gatherAssertTypes(__DIR__ . '/data/nullsafe-vs-scalar.php');
 		yield from $this->gatherAssertTypes(__DIR__ . '/data/bug-8517.php');
 	}
