+++ tests/PHPStan/Rules/Methods/NullsafeMethodCallRuleTest.php
@@ -4,6 +4,7 @@ namespace PHPStan\Rules\Methods;
 
 use PHPStan\Rules\Rule;
 use PHPStan\Testing\RuleTestCase;
+use const PHP_VERSION_ID;
 
 /**
  * @extends RuleTestCase<NullsafeMethodCallRule>
@@ -24,6 +25,25 @@ class NullsafeMethodCallRuleTest extends RuleTestCase
 				16,
 			],
 		]);
+	}
+
+	public function testNullsafeVsScalar(): void
+	{
+		 $this->analyse([__DIR__ . '/../../Analyser/data/nullsafe-vs-scalar.php'], []);
+	}
+
+	public function testBug8664(): void
+	{
+		$this->analyse([__DIR__ . '/../../Analyser/data/bug-8664.php'], []);
+	}
+
+	public function testBug9293(): void
+	{
+		if (PHP_VERSION_ID < 80000) {
+			$this->markTestSkipped('Test requires PHP 8.0.');
+		}
+
+		$this->analyse([__DIR__ . '/../../Analyser/data/bug-9293.php'], []);
 	}
 
 }
