+++ tests/PHPStan/Rules/Properties/NullsafePropertyFetchRuleTest.php
@@ -41,4 +41,32 @@ class NullsafePropertyFetchRuleTest extends RuleTestCase
 		$this->analyse([__DIR__ . '/data/bug-7109.php'], []);
 	}
 
+	public function testBug5172(): void
+	{
+		if (PHP_VERSION_ID < 80000) {
+			$this->markTestSkipped('Test requires PHP 8.0.');
+		}
+
+		$this->analyse([__DIR__ . '/../../Analyser/data/bug-5172.php'], []);
+	}
+
+	public function testBug7980(): void
+	{
+		if (PHP_VERSION_ID < 80100) {
+			$this->markTestSkipped('Test requires PHP 8.1.');
+		}
+
+		$this->analyse([__DIR__ . '/../../Analyser/data/bug-7980.php'], []);
+	}
+
+	public function testBug8517(): void
+	{
+		$this->analyse([__DIR__ . '/../../Analyser/data/bug-8517.php'], []);
+	}
+
+	public function testBug9105(): void
+	{
+		$this->analyse([__DIR__ . '/../../Analyser/data/bug-9105.php'], []);
+	}
+
 }
